+++ src/main/java/kitchenpos/domain/order/Order.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
 
 import java.time.LocalDateTime;
 import java.util.Objects;
@@ -11,6 +11,8 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.OneToOne;
 import javax.persistence.Table;
+
+import kitchenpos.domain.table.OrderTable;
 
 @Entity
 @Table(name = "orders")
