+++ src/main/java/kitchenpos/domain/product/Product.java
@@ -1,4 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.domain.product;
+
+import java.math.BigDecimal;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -7,7 +9,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
 @Entity
-public class MenuGroup {
+public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -16,11 +18,15 @@ public class MenuGroup {
     @Column
     private String name;
 
-    protected MenuGroup() {
+    @Column
+    private BigDecimal price;
+
+    protected Product() {
     }
 
-    public MenuGroup(final String name) {
+    public Product(final String name, final BigDecimal price) {
         this.name = name;
+        this.price = price;
     }
 
     public Long getId() {
@@ -29,5 +35,9 @@ public class MenuGroup {
 
     public String getName() {
         return name;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
     }
 }
