+++ src/main/java/kitchenpos/domain/table/OrderTables.java
@@ -1,8 +1,10 @@
-package kitchenpos.domain;
+package kitchenpos.domain.table;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
+
+import kitchenpos.domain.tablegroup.TableGroup;
 
 public class OrderTables {
