+++ src/main/java/kitchenpos/dto/menu/MenuProductResponse.java
@@ -3,7 +3,7 @@ package kitchenpos.dto.menu;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.domain.menu.MenuProduct;
 
 public class MenuProductResponse {
