+++ src/main/java/kitchenpos/dto/order/OrderLineItemResponse.java
@@ -3,7 +3,7 @@ package kitchenpos.dto.order;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.domain.order.OrderLineItem;
 
 public class OrderLineItemResponse {
