+++ src/main/java/kitchenpos/dto/table/OrderTableResponse.java
@@ -3,7 +3,7 @@ package kitchenpos.dto.table;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 
 public class OrderTableResponse {
