+++ src/main/java/kitchenpos/dto/tablegroup/TableGroupResponse.java
@@ -3,7 +3,7 @@ package kitchenpos.dto.tablegroup;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.tablegroup.TableGroup;
 import kitchenpos.dto.table.OrderTableResponse;
 
 public class TableGroupResponse {
