+++ src/main/java/kitchenpos/repository/MenuGroupRepository.java
@@ -3,7 +3,7 @@ package kitchenpos.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menugroup.MenuGroup;
 
 @Repository
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
