+++ src/main/java/kitchenpos/repository/MenuRepository.java
@@ -1,11 +1,9 @@
 package kitchenpos.repository;
 
-import java.util.List;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import kitchenpos.domain.Menu;
+import kitchenpos.domain.menu.Menu;
 
 @Repository
 public interface MenuRepository extends JpaRepository<Menu, Long> {
