+++ src/main/java/kitchenpos/repository/OrderLineItemRepository.java
@@ -1,9 +1,8 @@
 package kitchenpos.repository;
 
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.domain.order.OrderLineItem;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
