+++ src/main/java/kitchenpos/repository/OrderRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import kitchenpos.domain.Order;
+import kitchenpos.domain.order.Order;
 
 @Repository
 public interface OrderRepository extends JpaRepository<Order, Long> {
