+++ src/main/java/kitchenpos/repository/OrderTableRepository.java
@@ -5,8 +5,8 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.table.OrderTable;
+import kitchenpos.domain.tablegroup.TableGroup;
 
 @Repository
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
