+++ src/main/java/kitchenpos/repository/ProductRepository.java
@@ -1,9 +1,6 @@
 package kitchenpos.repository;
 
-import kitchenpos.domain.Product;
-
-import java.util.List;
-import java.util.Optional;
+import kitchenpos.domain.product.Product;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
