+++ src/main/java/kitchenpos/repository/TableGroupRepository.java
@@ -1,9 +1,6 @@
 package kitchenpos.repository;
 
-import kitchenpos.domain.TableGroup;
-
-import java.util.List;
-import java.util.Optional;
+import kitchenpos.domain.tablegroup.TableGroup;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
