+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -11,9 +11,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
 
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.table.OrderTable;
+import kitchenpos.domain.tablegroup.TableGroup;
 import kitchenpos.dto.table.OrderTableChangeEmptyRequest;
 import kitchenpos.dto.table.OrderTableChangeNumberOfGuestsRequest;
 import kitchenpos.dto.table.OrderTableRequest;
